django_users/requirements.txt
asgiref==3.11.0
django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
pyjwt==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.4
